Refactor ssh setup to a different plugin part
